Big update to the way attachment locations are considered in CASPR.
Instead of CableSegmentModel storing information regarding local and global attachment locations, new Attachment objects (with base class CableAttachmentBase) are used. These objects are responsible for storing the attachment information and the updating of them depending on the type of attachment.

Existing attachment type in the XML corresponds to the most commonly used CableAttachmentPoint. This allows new types of attachment points, for e.g. those that change as a function of pose, or those that are obtained from another program/library etc. Function to read the XML section for the attachments and to load the correct type of attachments is now placed within the CableModelBase (rather than the individual cable models). This allows different cable types and cable attachment types to be mixed and matched if needed.

Another important change is that r_OA (vector from origin to attachment) is now in the '0' INERTIAL frame and not in the local frame (this may affect other codes).